test: migrate to luatest
Rewrite existing regression and performance tests using luatest [1]
instead of tap. luatest is a modern tool to write tests for Tarantool
core and modules. It is expected for all new tests introduced to the
core to be written with luatest, so we rework this before embedding. Add
Makefile targets to install and clean dependencies.

1. https://github.com/tarantool/luatest

Part of tarantool/tarantool#7726