test: do not run on built-in
After embedding checks to the core binary, plain require('checks') will
return embedded module. So it won't be possible to run repository module
tests (in reusable_testing or with repository CI) with repository code
until it overloads build-in code. Until [1] is resolved there are no
any other way to deal with it. (And even after [1] would be resolved,
it may worth to leave it like this to support older versions with
built-in checks.) test/helper.lua is automatically required by luatest
before any run.

1. https://github.com/tarantool/tarantool/issues/7774

Part of tarantool/tarantool#7726